Cambios a NodeEntity, RoadRelationship, RoadRepository para que corra la pag

+++ src/main/java/com/example/demo/repository/RoadRepository.java
@@ -4,5 +4,5 @@ package com.example.demo.repository;
 import com.example.demo.model.RoadRelationship;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
-public interface RoadRepository extends Neo4jRepository<RoadRelationship, Integer> {
+public interface RoadRepository extends Neo4jRepository<RoadRelationship, Long> {
 }
